requirements.txt
fastapi
uvicorn
feedparser
requests
openai
python-dotenv
pytest
httpx
app